pyproject.toml
[project]
name = "nijimarutto-monitor"
version = "0.2.0"
description = "にじさんじショップ 在庫監視ツール"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "pyyaml>=6.0",
]

[project.scripts]
nijimarutto-monitor = "nijimarutto_monitor.monitor:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/nijimarutto_monitor"]
